% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLS.R
\name{IRLS}
\alias{IRLS}
\title{Function for iteratively reweighed least squares algorithm}
\usage{
IRLS(
  formula,
  data,
  threshold = 1e-06,
  state,
  t = NULL,
  a = NULL,
  b = NULL,
  c = NULL
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted, including the response and the predictors.The response is strictly restricted with rank one.}

\item{data}{the name of the data frame}

\item{threshold}{the threshold used to determine when the algorithm ends with default value 0.000001.}

\item{state}{the robust criterion function to apply, “Huber” represents Huber’s t function, “Ramsay” represents Ramsay’s Ea function, “Andrews” represents Andrews’ wave function, and “Hampel” represents Hampel’s 17A function.}

\item{t}{the value of t to be assigned to Huber’s t function if chosen}

\item{a}{the value of a to be assigned to Ramsay’s Ea function, Andrews’ wave function or Hampel’s 17A function if chosen.}

\item{b}{the value of b to be assigned to Hampel’s 17A function if chosen.}

\item{c}{the value of c to be assigned to Hampel’s 17A function if chosen.}
}
\value{
the calculated robust regression coefficient matrix
}
\description{
Function for iteratively reweighed least squares algorithm
}
\examples{
library(RARR)
IRLS(Petal.Width~Sepal.Length+Sepal.Width+Petal.Length, iris, state="Huber", t=2)
}
