% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_plot.R
\name{normal_plot}
\alias{normal_plot}
\title{Function for visualizing Normal Probability Plot}
\usage{
normal_plot(
  formula = "Petal.Width~Sepal.Length+Sepal.Width+Petal.Length",
  data = iris,
  resid_type = 1
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted, including the response and the predictors}

\item{data}{the name of the data frame}

\item{resid_type}{the type number of the scaling residuals, 1-Standardized Residuals, 2-Studentized Residuals, 3-Studentized PRESS Residuals, 4-R-student Residuals}
}
\value{
A plot showing whether the residuals follow normal distribution
}
\description{
Function for visualizing Normal Probability Plot
}
\examples{
library(RARR)
normal_plot()
}
