% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRESS_Statistics_and_Predicted_R2.R
\name{Press_and_R2}
\alias{Press_and_R2}
\title{Function for PRESS Statistics and Predicted R2}
\usage{
Press_and_R2(
  formula = "Petal.Width~Sepal.Length+Sepal.Width+Petal.Length",
  data = iris
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted, including the response and the predictors}

\item{data}{the name of the data frame}
}
\value{
the vector that contains the PRESS statistics and the predicted R2
}
\description{
Function for PRESS Statistics and Predicted R2
}
\examples{
library(RARR)
Press_and_R2()
}
