% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-Student_Residuals.R
\name{R_Student_Residuals}
\alias{R_Student_Residuals}
\title{Function for R-Student Residuals}
\usage{
R_Student_Residuals(
  formula = "Petal.Width~Sepal.Length+Sepal.Width+Petal.Length",
  data = iris
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted, including the response and the predictors}

\item{data}{the name of the data frame}
}
\value{
the R-Student Residuals of the fitted model
}
\description{
Function for R-Student Residuals
}
\examples{
library(RARR)
R_Student_Residuals()
}
