% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_regres_plot.R
\name{resid_regres_plot}
\alias{resid_regres_plot}
\title{Function to visualize scaling residuals versus Regressors}
\usage{
resid_regres_plot(
  formula = "Petal.Width~Sepal.Length+Sepal.Width+Petal.Length",
  data = iris,
  resid_type = 1
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted, including the response and the predictors}

\item{data}{the name of the data frame}

\item{resid_type}{the type number of the scaling residuals, 1-Standardized Residuals, 2-Studentized Residuals, 3-Studentized PRESS Residuals, 4-R-student Residuals}
}
\value{
a figure showing the relationship between residuals and regressors
}
\description{
Function to visualize scaling residuals versus Regressors
}
\examples{
library(RARR)
resid_regres_plot()
}
