% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robust_Influence.R
\name{Robust_influence}
\alias{Robust_influence}
\title{Function for robust influence}
\usage{
Robust_influence(z, criterion, t = NULL, a = NULL, b = NULL, c = NULL)
}
\arguments{
\item{z}{a scaled residual to be transformed into robust influence}

\item{criterion}{the robust criterion function to apply, “Huber” represents Huber’s t function, “Ramsay” represents Ramsay’s Ea function, “Andrews” represents Andrews’ wave function, and “Hampel” represents Hampel’s 17A function.}

\item{t}{the value of t to be assigned to Huber’s t function if chosen}

\item{a}{the value of a to be assigned to Ramsay’s Ea function, Andrews’ wave function or Hampel’s 17A function if chosen.}

\item{b}{the value of b to be assigned to Hampel’s 17A function if chosen.}

\item{c}{the value of c to be assigned to Hampel’s 17A function if chosen.}
}
\value{
the calculated robust influence function value of the input scaled residual
}
\description{
Function for robust influence
}
\examples{
library(RARR)
Robust_influence(0.2, "Huber", t=2)
}
