% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardized_Residuals.R
\name{Standardized_Residuals}
\alias{Standardized_Residuals}
\title{Function for Standardized Residuals}
\usage{
Standardized_Residuals(
  formula = "Petal.Width~Sepal.Length+Sepal.Width+Petal.Length",
  data = iris
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted, including the response and the predictors}

\item{data}{the name of the data frame}
}
\value{
the standardized residuals of the fitted model
}
\description{
Function for Standardized Residuals
}
\examples{
library(RARR)
Standardized_Residuals()
}
