% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Studentized_PRESS_Residuals.R
\name{Studentized_Press_Residuals}
\alias{Studentized_Press_Residuals}
\title{Function for PRESS Statistics and Predicted R2}
\usage{
Studentized_Press_Residuals(
  formula = "Petal.Width~Sepal.Length+Sepal.Width+Petal.Length",
  data = iris
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted, including the response and the predictors}

\item{data}{the name of the data frame}
}
\value{
the Studentized PRESS Residuals of the fitted model
}
\description{
Function for PRESS Statistics and Predicted R2
}
\examples{
library(RARR)
Studentized_Press_Residuals()
}
