% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Studentized_Residuals.R
\name{Studentized_Residuals}
\alias{Studentized_Residuals}
\title{Function for Studentized Residuals}
\usage{
Studentized_Residuals(
  formula = "Petal.Width~Sepal.Length+Sepal.Width+Petal.Length",
  data = iris
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted, including the response and the predictors}

\item{data}{the name of the data frame}
}
\value{
the studentized residuals of the fitted model
}
\description{
Function for Studentized Residuals
}
\examples{
library(RARR)
Studentized_Residuals()
}
